requirements.txt
# Web framework
fastapi>=0.104.1
uvicorn>=0.24.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI and ML
openai>=1.3.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.3
tiktoken>=0.6.0
transformers>=4.36.0

# Data processing
numpy>=1.25.0
pandas>=2.0.0

# Vector search
faiss-cpu>=1.7.4

# Utilities
pydantic>=2.5.1
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.8

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1

# Voice interface
pyttsx3>=2.90
SpeechRecognition>=3.10.0
chromadb>=0.5.0

